# [TT-12702] revert wrappedServeHTTP to use recordDetail