--- a/.gitignore
+++ b/.gitignore
@@ -61,3 +61,5 @@ tyk_linux_*
 *.test
 
 main
+
+/coprocess/*.pb.go-e