@lghiur 2024-10-23
API tests failure, and Otel e2e  failures are not related to this PR